requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
Automat==0.7.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cryptography-vectors==2.3.1
cssselect==1.0.3
decorator==4.3.0
fake-useragent==0.1.10
hyperlink==17.3.1
idna==2.7
incremental==17.5.0
ipython-genutils==0.2.0
isort==4.3.4
jsonschema==2.6.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.2.0
mysql-connector==2.1.6
nbformat==4.4.0
numpy==1.15.1
parsel==1.5.0
plotly==3.1.1
pluggy==0.7.1
prettytable==0.7.2
py==1.5.4
pyasn1==0.4.4
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.1.1
pyOpenSSL==18.0.0
pytest==3.7.1
pytest-runner==4.2
pytz==2018.5
pywin32==221
queuelib==1.5.0
requests==2.19.1
retrying==1.3.3
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.5.1
scrapy-fake-useragent==1.1.0
scrapy-proxies==0.4
Scrapy-UserAgents==0.0.1
service-identity==17.0.0
six==1.11.0
traitlets==4.3.2
Twisted==18.7.0
typed-ast==1.1.0
urllib3==1.23
w3lib==1.19.0
wincertstore==0.2
wrapt==1.10.11
zope.interface==4.5.0
